LayoutTests/fast/images/text-recognition/image-overlay-block-with-newlines-expected.txt

PASS getSelection().toString() is "Hello\nworld"
PASS successfullyParsed is true

TEST COMPLETE
